+++ cargo-crev/CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Ignore cargo directory source replacements - this essentially ignores `cargo vendor` and `cargo crev` will not read vendored sources
 - Proofs moved from the config directory (e.g. ~/.config/crev/proofs)
   to the data directory (e.g. ~/.local/share/crev/proofs).
+- Added command `config data-dir`.
+- Added command `config cache-dir`.
 
 ## [0.22.2](https://github.com/dpc/crev/compare/cargo-crev-v0.21.4...v0.22.2) - 2022-01-11